Major refactor, using inheritance to share code
