+++ dreimac/__init__.py
@@ -1,5 +1,5 @@
 from .circularcoords import CircularCoords
 from .projectivecoords import ProjectiveCoords
-from .geomtools import PartUnity, GeometryExamples, get_csm, get_csm_projarc, get_ssm, get_greedy_perm_pc, get_greedy_perm_dm
+from .utils import GeometryUtils, PartUnity, GeometryExamples, PlotUtils
 
 from ._version import __version__
